+++ CHANGES.md
@@ -1,3 +1,6 @@
+# 0.?.? - 202?-??-??
+- CritiqueBrainz links for recordings
+
 # 0.8.0 - 2022-12-12
 - Show link to release group on CritiqueBrainz in release onebox (#17)
 - CritiqueBrainz links for labels and works (#17)